+++ chrome-cordova/chrome-apps-api-tests/tests/auto/test.chrome.gcm.js
@@ -4,12 +4,10 @@
 
 registerAutoTests("chrome.gcm", function() {
   'use strict';
-  var testTimeout = 2000;
   var senderid = '90031296475';
-//  var senderid = '594483801284';
   var sender = senderid+ "@gcm.googleapis.com";
 
-  describeAndroidOnly('testing registration', function() {
+  describeExcludeIos('registration', function() {
     it('should contain definitions', function(done) {
       expect(chrome.gcm).toBeDefined();
       expect(chrome.gcm.send).toBeDefined();
@@ -23,6 +21,7 @@ registerAutoTests("chrome.gcm", function() {
 
     it('should register and unregister', function(done) {
       chrome.gcm.register([senderid], function(regid) {
+        expect(regid).toBeDefined();
         expect(regid.length).toBeGreaterThan(1);
         expect(chrome.runtime.lastError).not.toBeDefined();
         chrome.gcm.unregister(function() {
@@ -38,7 +37,7 @@ registerAutoTests("chrome.gcm", function() {
           expect('Not to get here').toEqual('');
         });
       } catch(e) {
-        expect(e.message.substring(0,13)).toEqual('Invalid value');
+        expect(e.message).toBeDefined();
         expect(chrome.runtime.lastError).not.toBeDefined();
       }
       done();
@@ -52,6 +51,24 @@ registerAutoTests("chrome.gcm", function() {
       });
     });
 
+  });
+
+  describeExcludeIos('sending', function() {
+    it('should error for missing key', function(done) {
+      var message = {
+        'data' : { 'test' : 'test' }
+      };
+      try {
+        chrome.gcm.send(message, function(msgid) {
+          expect("Should not be here").not.toBeDefined();
+          done();
+        });
+      } catch (e) {
+        expect(e.message).toBeDefined();
+        done();
+      }
+    });
+
     it('should error for invalid data', function(done) {
       var message = {
         'messageId' : '0',
@@ -61,59 +78,21 @@ registerAutoTests("chrome.gcm", function() {
       };
       try {
         chrome.gcm.send(message, function(msgid) {
-          expect(msgid.length).toBeGreaterThan(0);
+          expect("Should not be here").not.toBeDefined();
           done();
         });
       } catch (e) {
-          expect(e.message).toEqual("Invalid data key: collapse_key");
-          done();
-      }
-    });
-  });
-
-  describeAndroidOnly('testing sending', function() {
-
-    var msgcount=0;
-    var globdone=null;
-    var listener = function(msg) {
-      console.log(' msg',msg);
-      msgcount++;
-      if(globdone) globdone();
-    }
-    beforeEach( function() {
-      chrome.gcm.onMessage.addListener(listener);
-    })
-
-    afterEach( function() {
-      chrome.gcm.onMessage.removeListener(listener);
-      expect(msgcount=1);
-    })
-
-    it('should send and receive one msg', function(done) {
-      globdone=done;
-      var message = {
-        'destinationId' : sender,
-        'messageId' : '0',
-        'timeToLive' : 10,
-        'data' : { 'test' : 'test' }
-      };
-      try {
-        chrome.gcm.send(message, function(msgid) {
-          expect(msgid.length).toBeGreaterThan(0);
-          expect(chrome.runtime.lastError).not.toBeDefined();
-        });
-      } catch (e) {
-          expect(e).not.toBeDefined();
+        expect(e.message).toBeDefined();
+        done();
       }
     });
 
     it('should fail to send a big msg', function(done) {
-      globdone=done;
-      var blob100='0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789';
+      var blob100 = '0123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890123456789';
       var msgdata = {};
-      for(var k=0;k<41;k++){
-        var key='k'+k;
-        msgdata[key]=blob100;
+      for(var k = 0; k < 41; k++){
+        var key = 'k' + k;
+        msgdata[key] = blob100;
       }
       var message = {
         'destinationId' : sender,
@@ -123,13 +102,81 @@ registerAutoTests("chrome.gcm", function() {
       };
       try {
         chrome.gcm.send(message, function(msgid) {
-          expect('not to get there');
+          expect('Should not get here').toBe(false);
+          done();
         });
       } catch (e) {
-          expect(e.message).toBeDefined();
-          expect(e.message.substring(0,16)).toEqual("Payload exceeded");
+        expect(e.message).toBeDefined();
+        //expect(e.message.substring(0,16)).toEqual("Payload exceeded");
+        done();
       }
-      done();
+    });
+
+    it('should send and receive one simple msg', function(done) {
+      var message = {
+        'destinationId' : sender,
+        'messageId' : '0',
+        'timeToLive' : 10,
+        'data' : { 'type': 'ping', 'message' : 'test' }
+      };
+      chrome.gcm.onMessage.addListener(function listener(msg) {
+        expect(msg).toBeDefined();
+        expect(msg.data).toBeDefined();
+        expect(msg.data.type).toEqual('pong')
+        expect(msg.data.message).toEqual('test')
+
+        chrome.gcm.onMessage.removeListener(listener);
+        done();
+      });
+      try {
+        chrome.gcm.send(message, function(msgid) {
+          expect(msgid).toBeDefined();
+          expect(msgid.length).toBeGreaterThan(0);
+          expect(chrome.runtime.lastError).not.toBeDefined();
+        });
+      } catch (e) {
+          expect(e).not.toBeDefined();
+      }
+    });
+
+    it('should send and receive one complex msg', function(done) {
+      var message = {
+        'destinationId' : sender,
+        'messageId' : '0',
+        'timeToLive' : 10,
+        'data' : {
+          'type': 'ping',
+          'message' : 'testing multi value',
+          'message2': 'more2',
+          'message3': 'more3',
+          'message4': 'more4',
+          'message5': 'more5',
+          //'int': 1,
+          //'float': 3.33,
+          //'bool': true,
+          //'string': 'sss',
+          //'array': [1, 3.33, true, 'sss', [1,2,3], {'a':1, 'b':2}],
+          //'map': {'a': 1, 'b': 2}
+        }
+      };
+      chrome.gcm.onMessage.addListener(function listener(msg) {
+        expect(msg).toBeDefined();
+        expect(msg.data).toBeDefined();
+        expect(msg.data.type).toEqual('pong')
+        expect(msg.data.message).toEqual(message.data.message)
+
+        chrome.gcm.onMessage.removeListener(listener);
+        done();
+      });
+      try {
+        chrome.gcm.send(message, function(msgid) {
+          expect(msgid).toBeDefined();
+          expect(msgid.length).toBeGreaterThan(0);
+          expect(chrome.runtime.lastError).not.toBeDefined();
+        });
+      } catch (e) {
+          expect(e).not.toBeDefined();
+      }
     });
 
     // I would love to test onSendError and onMessagesDeleted