+++ src/create-app.js
@@ -2,6 +2,8 @@ var path = require('path');
 var Q = require('q');
 var fs = require('fs');
 var shelljs = require('shelljs');
+var xmldom = require('xmldom');
+var ccaManifestLogic = require('cca-manifest-logic');
 
 function resolveTilde(string) {
   // TODO: implement better
@@ -70,8 +72,7 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
     }
     manifest = manifestData;
   })
-  .then(require('./tools-check'))
-  .then(function(toolsCheckResults) {
+  .then(function() {
     // Create step.
     console.log('## Creating Your Application');
     var config_default = JSON.parse(JSON.stringify(require('./default-config')(ccaRoot)));
@@ -87,9 +88,10 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
     if (!appWasImported) {
       // Update app name if the app is not imported.
       return Q.ninvoke(fs, 'readFile', manifestDesktopFilename, { encoding: 'utf-8' }).then(function(manifestDesktopData) {
+        var manifestDesktop;
         try {
           // jshint evil:true
-          var manifestDesktop = eval('(' + manifestDesktopData + ')');
+          manifestDesktop = eval('(' + manifestDesktopData + ')');
           // jshint evil:false
         } catch (e) {
           console.error(e);
@@ -98,7 +100,7 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
         manifestDesktop.name = appName || path.basename(destAppDir);
         manifest.name = manifestDesktop.name;
         Q.ninvoke(fs, 'writeFile', manifestDesktopFilename, JSON.stringify(manifestDesktop, null, 4));
-      })
+      });
     }
   })
   .then(function() {
@@ -111,6 +113,7 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
   .then(function() {
     // Update default packageId if needed.
     return Q.ninvoke(fs, 'readFile', manifestMobileFilename, { encoding: 'utf-8' }).then(function(manifestMobileData) {
+      var manifestMobile;
       try {
         // jshint evil:true
         manifestMobile = eval('(' + manifestMobileData + ')');
@@ -123,7 +126,7 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
         manifestMobile.packageId = packageId || ('com.your.company.' + (appName || manifest['name'].replace(/[^a-zA-Z0-9_]/g, '')));
         Q.ninvoke(fs, 'writeFile', manifestMobileFilename, JSON.stringify(manifestMobile, null, 4));
       }
-    })
+    });
   })
   .then(function() {
     // If there is no config.xml, or the config.xml is the cordova default, replace it with our default
@@ -133,9 +136,15 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
     }
   })
   .then(function() {
-    return require('./update-config-xml')();
+    return require('./get-manifest')('www');
   })
-  .then(function() {
+  .then(function(manifestJson) {
+    var configXmlData = fs.readFileSync('config.xml', 'utf8');
+    var analyzedManifest = ccaManifestLogic.analyseManifest(manifestJson);
+    var configXmlDom = new xmldom.DOMParser().parseFromString(configXmlData);
+    ccaManifestLogic.updateConfigXml(manifestJson, analyzedManifest, configXmlDom);
+    configXmlData = new xmldom.XMLSerializer().serializeToString(configXmlDom);
+    fs.writeFileSync('config.xml', configXmlData);
     return require('./write-out-cca-version')();
   })
   .then(function() {
@@ -198,7 +207,7 @@ module.exports = exports = function createApp(destAppDir, ccaRoot, origDir, pack
       welcomeText += 'Your project has been created, with web assets in the `www` directory:\n'+
                      wwwPath + '\n\n';
     }
-    welcomeText += 'Remember to run `cca prepare` after making changes\n';
+    welcomeText += 'Remember to run `cca prepare` after making changes if you are using an IDE.\n';
     welcomeText += 'Full instructions: https://github.com/MobileChromeApps/mobile-chrome-apps/blob/master/docs/Develop.md#making-changes-to-your-app-source-code';
     console.log(welcomeText);
   });