+++ src/get-manifest.js
@@ -17,46 +17,16 @@
   under the License.
  */
 
-var Q = require('q');
-var path = require('path');
 var fs = require('fs');
+var path = require('path');
+var Q = require('q');
+var ccaManifestLogic = require('cca-manifest-logic');
 
 // Returns a promise for the manifest contents.
 module.exports = exports = function getManifest(manifestDir, platform) {
   var manifestFilename = path.join(manifestDir, 'manifest.json');
-  var manifestMobileFilename = path.join(manifestDir, 'manifest.mobile.json');
-  var manifest, manifestMobile;
-
-  return Q.ninvoke(fs, 'readFile', manifestFilename, { encoding: 'utf-8' }).then(function(manifestData) {
-    try {
-      // jshint evil:true
-      manifest = eval('(' + manifestData + ')');
-      // jshint evil:false
-    } catch (e) {
-      console.error(e);
-      return Q.reject('Unable to parse manifest ' + manifestFilename);
-    }
-    return Q.ninvoke(fs, 'readFile', manifestMobileFilename, { encoding: 'utf-8' }).then(function(manifestMobileData) {
-      try {
-        // jshint evil:true
-        manifestMobile = eval('(' + manifestMobileData + ')');
-        // jshint evil:false
-      } catch (e) {
-        console.error(e);
-        return Q.reject('Unable to parse manifest ' + manifestMobileFilename);
-      }
-    }, function(err) {
-      // Swallow any errors opening the mobile manifest, it's not required.
-    }).then(function() {
-      var extend = require('node.extend');
-      manifest = extend(true, manifest, manifestMobile); // true -> deep recursive merge of these objects
-      // If you want a specific platform manifest, also merge in its platform specific settings
-      if (typeof platform !== 'undefined' && platform in manifest) {
-        manifest = extend(true, manifest, manifest[platform]);
-      }
-      return manifest;
-    });
-  }, function(err) {
-    return Q.reject('Unable to open manifest ' + manifestFilename + ' for reading.');
-  });
+  function readFileFunc(p) {
+    return Q.ninvoke(fs, 'readFile', p, 'utf8');
+  }
+  return ccaManifestLogic.parseAndMergeManifests(manifestFilename, platform, readFileFunc, Q);
 };