+++ src/push-to-harness.js
@@ -21,7 +21,8 @@ module.exports = exports = function push(target, watch) {
     return ret.then(function(targets) {
       return createSession(targets);
     }).then(function(session) {
-      return pushAll(session.clientInfos)
+      // Push the app and force a launch, even if it hasn't changed from a previous launch, since the app is being pushed for the first time.
+      return pushAll(session.clientInfos, /* forceLaunch */ true)
       .then(function() {
         if (watch) {
           return watchFiles(session);
@@ -104,7 +105,11 @@ function createSession(targets) {
 
 var pushInProgress = false;
 var pushAgainWhenDone = false;
-function pushAll(clientInfos) {
+// gaze uses fs.watchFile, which is classified as "unstable" (http://goo.gl/H16j5l).
+// It sometimes causes multiple change events to be fired, and they're far enough apart that debouncing is ineffective.
+// To deal with this, CADT-client's push functionality doesn't relaunch an app with no changes unless `forceLaunch` is true.
+// Calls to `pushAll` can specify whether to force a launch.
+function pushAll(clientInfos, forceLaunch) {
   if (pushInProgress) {
     pushAgainWhenDone = true;
     return Q();
@@ -112,7 +117,7 @@ function pushAll(clientInfos) {
   pushInProgress = true;
   var allPromises = clientInfos.map(function(clientInfo) {
     try {
-      return clientInfo.pushSession.push();
+      return clientInfo.pushSession.push(forceLaunch);
     } catch (e) {
       if (/Not a Cordova project/.test(e)) {
         console.warn('Please navigate to a Chrome App or Cordova project, and then try pushing again.');
@@ -126,7 +131,8 @@ function pushAll(clientInfos) {
   .then(function() {
     if (pushAgainWhenDone) {
       process.nextTick(function() {
-        pushAll(clientInfos).done();
+        // Push, but don't force a launch if the app hasn't changed.
+        pushAll(clientInfos, /* forceLaunch */ false).done();
       });
     }
     pushInProgress = false;
@@ -139,7 +145,8 @@ function pushAll(clientInfos) {
 }
 
 var debouncedPushAll = debounce(function(clientInfos) {
-  pushAll(clientInfos).done();
+  // Push, but don't force a launch if the app hasn't changed.
+  pushAll(clientInfos, /* forceLaunch */ false).done();
 }, 50, false);
 
 function watchFiles(session) {